update v2.0.3
all spawned crops are now dead
trees now have a 20% chance to drop a seed (no more guaranteed seeds from trees)
Well Rested buff now gives stamina regen in stages instead of a flat buff
Lowered chance of "fast" zombies in random movement types
Despair will now treat faster if the player is on a bedroll next to a campfire
Fixed a bug with liquid nitrogen throwing NRE on enemy hit
Changed the Iron Hoe to a Fertilizer Bucket. Works the same as before, just a different item.
Fixed rarity of canned carrots
Swapped some trader spawns. Bob is now in the burnt forest, Jen in the desert, Joel in the snow and Hugh in the Wasteland. Traders Red & Hells can now be found in the wasteland and desert respectively
Edited health bars so all information is now visible if a health bar mod is used
Fixed scrapping weight of disposable battery
Localization fixes and tweaks.